% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_feature}
\alias{plot_feature}
\title{Plot tSNE with continuous variable overlaid}
\usage{
plot_feature(seurat_obj, ident = NULL, gene = NULL, plot_dat = NULL,
  plot_col = NULL, meta_data_col = NULL, pt.size = 0.001,
  pt.alpha = 0.25, label_text = FALSE, label.size = 6, .cols = NULL,
  legend_names = NULL, cell_filter = NULL, palette_type = "cloupe",
  col_pal = "Reds", max_y = NULL, ...)
}
\description{
Plot tSNE with continuous variable overlaid
}
