% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph}
\alias{plot_graph}
\title{Calculate force directed graph coordinates}
\usage{
plot_graph(graph_df, color_by = "orig.ident")
}
\arguments{
\item{graph_df}{data.frame produced by calc_graph}

\item{color_by}{column from graph_df to color cells by}
}
\value{
ggplot object
}
\description{
Calculate force directed graph coordinates
}
