% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellar.R
\docType{package}
\name{cellar}
\alias{cellar}
\alias{cellar-package}
\title{cellar: functions and pipelines for processing single cell data}
\description{
cellar provides a collection of snakemake pipelines and analysis vignettes
             for single cell RNA-Seq data.
}
\seealso{
Report bugs at \url{https://github.com/rnabioco/cellar}
}
\author{
Kent Riemondy \href{mailto:kent.riemondy@gmail.com}{kent.riemondy@gmail.com}
}
